chore: organize loaders
