+++ client/src/features/loaders/index.ts
@@ -1,0 +1,2 @@
+export { default as ContentLoader } from "./ContentLoader";
+export { default as AppLoader } from "./AppLoader";
